+++ view/DataList.go
@@ -2,5 +2,5 @@ package view
 
 type DataList struct {
 	Items interface{} `form:"items" json:"items"`
-	Total int         `form:"total" json:"total"`
+	Total int64       `form:"total" json:"total"`
 }
